refactor: 로그 도메인을 각 소유 도메인 하위로 이동
domain/log 통합 디렉토리를 제거하고 MemberStatusLog/MemberGradeLog는
member/log, SoftwareStatusLog는 software/log 하위로 재배치해 도메인 응집도를 높임.
