+++ src/main/java/koza/licensemanagementservice/domain/member/controller/MemberAdminController.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import koza.licensemanagementservice.auth.dto.CustomUser;
-import koza.licensemanagementservice.domain.log.dto.response.MemberGradeLogResponse;
-import koza.licensemanagementservice.domain.log.dto.response.MemberStatusLogResponse;
+import koza.licensemanagementservice.domain.member.log.dto.response.MemberGradeLogResponse;
+import koza.licensemanagementservice.domain.member.log.dto.response.MemberStatusLogResponse;
 import koza.licensemanagementservice.domain.member.dto.request.MemberGradeChangeRequest;
 import koza.licensemanagementservice.domain.member.dto.request.MemberStatusChangeRequest;
 import koza.licensemanagementservice.domain.member.service.MemberAdminService;
