+++ src/main/java/koza/licensemanagementservice/domain/member/log/repository/MemberStatusLogRepository.java
@@ -1,6 +1,6 @@
-package koza.licensemanagementservice.domain.log.repository;
+package koza.licensemanagementservice.domain.member.log.repository;
 
-import koza.licensemanagementservice.domain.log.entity.MemberStatusLog;
+import koza.licensemanagementservice.domain.member.log.entity.MemberStatusLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
