+++ src/main/java/koza/licensemanagementservice/domain/software/log/entity/SoftwareStatusLog.java
@@ -1,4 +1,4 @@
-package koza.licensemanagementservice.domain.log.entity;
+package koza.licensemanagementservice.domain.software.log.entity;
 
 import jakarta.persistence.*;
 import koza.licensemanagementservice.domain.member.entity.Member;
